Move format loaders into query-graphs/src/loaders/
I am planning to add a couple more loaders. In preparation for that,
move all loaders into the src/loaders/ subdirectory, so they are
separated from other files.

Also fix query-graphs/package.json's `files` glob (`lib/*.js` ->
`lib/**/*.js`), which only matched top-level lib/ files and silently
dropped lib/ui/ and now lib/loaders/ from the published npm package.

`assert`/`assertNotNull` were moved out of loader-utils.ts into their
own top-level `assert.ts`: they are generic, loader-independent
utilities.
