+++ query-graphs/package.json
@@ -28,8 +28,8 @@
   ],
   "license": "MIT",
   "files": [
-    "lib/*.js",
-    "lib/*.d.ts",
+    "lib/**/*.js",
+    "lib/**/*.d.ts",
     "style/*.*"
   ],
   "directories": {
