+++ query-graphs/src/loaders/hyper.ts
@@ -20,8 +20,8 @@ We transform a Hyper JSON tree into a query-graphs tree using the following heur
 
 */
 
-import type {TreeNode, TreeDescription, Crosslink, IconName} from "./tree-description";
-import {allChildren} from "./tree-description";
+import type {TreeNode, TreeDescription, Crosslink, IconName} from "../tree-description";
+import {allChildren} from "../tree-description";
 import type {Json, JsonObject} from "./loader-utils";
 import {forceToString, tryToString, formatMetric, hasOwnProperty, tryGetPropertyPath} from "./loader-utils";
