+++ query-graphs/src/loaders/json.ts
@@ -9,7 +9,7 @@ Map the JSON tree directly to a D3 tree, without any modifications
 
 import type {Json} from "./loader-utils";
 import {tryToString} from "./loader-utils";
-import type {TreeDescription, TreeNode} from "./tree-description";
+import type {TreeDescription, TreeNode} from "../tree-description";
 
 function convertChildren(node: Json): TreeNode[] {
     const strRep = tryToString(node);
