+++ query-graphs/src/loaders/loader-utils.ts
@@ -84,16 +84,3 @@ export function formatMetric(x: number): string {
     }
     return x.toFixed(0) + sizes[idx];
 }
-
-export function assert(value: boolean, errorMsg = "Assertion violated"): asserts value {
-    if (!value) {
-        // eslint-disable-next-line no-debugger
-        debugger;
-        throw new Error(errorMsg);
-    }
-}
-
-export function assertNotNull<T>(v: T | null | undefined): asserts v is T {
-    assert(v !== null, "Unexpected null value");
-    assert(v !== undefined, "Unexpected undefined value");
-}
