+++ query-graphs/src/loaders/postgres.ts
@@ -7,10 +7,11 @@ This is pretty much the same algorithm as the algorithm for Hyper plans
 
 */
 
-import * as treeDescription from "./tree-description";
-import type {TreeNode, TreeDescription, Crosslink, IconName} from "./tree-description";
+import * as treeDescription from "../tree-description";
+import type {TreeNode, TreeDescription, Crosslink, IconName} from "../tree-description";
 import type {Json} from "./loader-utils";
-import {assert, tryToString, formatMetric, hasOwnProperty, hasSubOject} from "./loader-utils";
+import {tryToString, formatMetric, hasOwnProperty, hasSubOject} from "./loader-utils";
+import {assert} from "../assert";
 
 interface UnresolvedCrosslink {
     source: TreeNode;
